1581 step up activity component
