+++ core/src/main/java/org/kiwix/kiwixmobile/core/utils/AlertDialogShower.kt
@@ -21,11 +21,14 @@ package org.kiwix.kiwixmobile.core.utils
 import android.app.Activity
 import androidx.appcompat.app.AlertDialog
 import org.kiwix.kiwixmobile.core.R
+import org.kiwix.kiwixmobile.core.di.ActivityScope
 import org.kiwix.kiwixmobile.core.utils.KiwixDialog.StartHotspotManually
 import javax.inject.Inject
+import javax.inject.Named
 
+@ActivityScope
 class AlertDialogShower @Inject constructor(
-  private val activity: Activity,
+  @Named("activity") private val activity: Activity,
   private val sharedPreferenceUtil: SharedPreferenceUtil
 ) : DialogShower {
   override fun show(
